[#42] add search by programming language to Communities#index (#44)
Add DatabaseCleaner to both profile and user spec to clear out DB between tests.
Moved mentee/mentor to Profile since Communities#index was primarily returning profiles
and because the search was on Profile, not User. Added pg_search gem to allow for text
search since, currently, languages are a text field, not a drop down selection.